Temporary patch for bug 1358192: printf("%f"...) sets fraction to zero.
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@3950 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ device/lib/printf_large.c
@@ -211,7 +211,7 @@ static int output_float (float f, unsigned char reqWidth,
 
   // split the float
   integerPart=f;
-  decimalPart=f-integerPart;
+  decimalPart=f-(float)integerPart;
 
   // fill the buffer with the integerPart (in reversed order!)
   while (integerPart) {
@@ -235,7 +235,7 @@ static int output_float (float f, unsigned char reqWidth,
       // truncate the float
       integerPart=decimalPart;
       fpBuffer[fpBD++]='0' + integerPart;
-      decimalPart-=integerPart;
+      decimalPart-=(float)integerPart;
     } while (--i);
   }
